Extract the isolated server lifecycle into its own module
startServer, makeStop, responds, ENVIRONMENT_FOR and the port/temp-root
constants move out of site-characterization-rebaseline.mjs into a new
scripts/isolated-server.mjs, which imports them back. A second caller is
coming — a per-environment `--check` wrapper — and the piece being shared
is the isolation logic, whose failure mode is corrupting a running
server's resources/_gen. One home for it, rather than a home and a copy.

Deliberately not folded into dev-server.mjs, which answers a different
question (what server should the harness use, including reusing one it
finds and building into the repo's own docs/) and is imported by
site-characterization.mjs, smoke-pages.mjs and
site-characterization-probe-control.mjs.

Three changes beyond the move, each undone in the proof below:
ISO_ROOT's temp directory is renamed sc-rebaseline -> sc-isolated, since
two callers now write there; five declarations gain `export`; and the
preflight's "Add them to ENVIRONMENT_FOR in this file" now names
scripts/isolated-server.mjs.

Proved a pure relocation by reverse transform: the pre-move file rebuilt
from the two current ones, undoing only those three normalizations plus
the import lines, diffs byte-identical against a snapshot of the original
-- 629 lines both sides, diff exit 0. That is what caught a leaked
`const DISCARD` and an orphaned DIRTY_LINES comment, both of which
`node --check` accepted. Also: node --check on both files, an import()
listing exactly the five intended exports, a symbol sweep putting spawn /
tmpdir / sleep / HUGO_BIN / PREFIXES / makeStop / SERVER_TIMEOUT_S at zero
occurrences in rebaseline and non-zero in the module, and --help running
clean through the new import.

startServer itself was exercised end to end: it brought up an isolated
prod_prod server on :8090 serving /IndicatorPublic/, with HUGO_RESOURCEDIR
and -d under the temp root, and the harness selected the prod_prod
baseline unaided and enumerated 925 pages.
